# server/docs/swagger.yaml
basePath: /
definitions:
...
  apidocs.BootstrapRequest:
    properties:
      display_name:
        example: Super Admin
        type: string
      email:
        example: admin@tofex.com
        type: string
      password:
        example: changeme1234
        type: string
      username:
        example: superadmin
        type: string
    type: object
  apidocs.BootstrapResponse:
    properties:
      message:
        example: first admin created
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  apidocs.ErrorBody:
    properties:
      code:
        example: validation_error
        type: string
      message:
        example: invalid request
        type: string
    type: object
  apidocs.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apidocs.ErrorBody'
    type: object
  apidocs.ForgotPasswordRequest:
    properties:
      email:
        example: admin@tofex.com
        type: string
    type: object
...
  apidocs.LoginRequest:
    properties:
      login:
        example: admin@tofex.com
        type: string
      password:
        example: changeme1234
        type: string
    type: object
  apidocs.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        example: 86400
        type: integer
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  apidocs.MessageResponse:
    properties:
      message:
        example: ok
        type: string
    type: object
...
  apidocs.PublicCreateOrderResponse:
    properties:
      order_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        example: pending
        type: string
      total_cents:
        example: 900000
        type: integer
    type: object
...
  apidocs.ResetPasswordRequest:
    properties:
      new_password:
        example: newpassword1234
        type: string
      token:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      label:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.RolePermission'
        type: array
      updated_at:
        type: string
    type: object
  models.RolePermission:
    properties:
      permission:
        type: string
      role_id:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for Tofex restaurant/confectionery chain (staff admin + public
    guest ordering).
  title: Tofex API
  version: "1.0"
paths:
...
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Generic response; dev mode may include reset_token when LOG_PASSWORD_RESET_LINK=true
          schema:
            $ref: '#/definitions/apidocs.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      summary: Forgot password
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials (email or username)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      summary: Staff login
      tags:
      - auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset token + new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      summary: Reset password
      tags:
      - auth
  /api/v1/bootstrap:
    post:
      consumes:
      - application/json
      description: One-time setup. Disabled after any user exists.
      parameters:
      - description: First admin
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.BootstrapRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.BootstrapResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      summary: Bootstrap first admin
      tags:
      - auth
  /api/v1/public/orders:
    post:
      consumes:
      - multipart/form-data
      description: Anonymous checkout. Requires receipt image and chosen payment account.
        items is JSON array e.g. [{"product_id":1,"quantity":2}].
      parameters:
      - description: Customer full name
        in: formData
        name: customer_name
        required: true
        type: string
      - description: Customer email
        in: formData
        name: customer_email
        required: true
        type: string
      - description: Customer phone
        in: formData
        name: customer_phone
        type: string
      - description: Optional note
        in: formData
        name: customer_note
        type: string
      - description: Active payment account ID
        in: formData
        name: payment_account_id
        required: true
        type: integer
      - description: JSON array of line items
        in: formData
        name: items
        required: true
        type: string
      - description: Payment receipt image
        in: formData
        name: receipt
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.PublicCreateOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      summary: Create guest order (public)
      tags:
      - public
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"